requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
attrs==25.3.0
    # via pytest
black==25.1.0
    # via -r requirements.in
click==8.2.1
    # via black
coverage[toml]==7.9.1
    # via pytest-cov
django==5.2.3
    # via
    #   -r requirements.in
    #   django-model-utils
    #   djangorestframework
    #   model-bakery
django-environ==0.9.0
    # via -r requirements.in
django-model-utils==5.0.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via -r requirements.in
iniconfig==2.1.0
    # via pytest
model-bakery==1.20.5
    # via -r requirements.in
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements.in
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
sqlparse==0.5.3
    # via django
